Remove collectors and keymanager
