--- src/ert/_c_wrappers/enkf/export/__init__.py
@@ -1,19 +1,0 @@
-from .arg_loader import ArgLoader
-from .design_matrix_reader import DesignMatrixReader
-from .gen_data_collector import GenDataCollector
-from .gen_data_observation_collector import GenDataObservationCollector
-from .gen_kw_collector import GenKwCollector
-from .misfit_collector import MisfitCollector
-from .summary_collector import SummaryCollector
-from .summary_observation_collector import SummaryObservationCollector
-
-__all__ = [
-    "DesignMatrixReader",
-    "SummaryCollector",
-    "SummaryObservationCollector",
-    "GenKwCollector",
-    "MisfitCollector",
-    "GenDataCollector",
-    "GenDataObservationCollector",
-    "ArgLoader",
-]
